+++ libs/checkpoint-postgres/langgraph/store/postgres/aio_store.py
@@ -7,13 +7,13 @@ from psycopg import AsyncConnection, AsyncCursor
 from psycopg.errors import UndefinedTable
 from psycopg.rows import dict_row
 
-from langgraph.checkpoint.postgres.base_store import (
+from langgraph.store.base import GetOp, ListNamespacesOp, Op, PutOp, Result, SearchOp
+from langgraph.store.postgres.base_store import (
     BasePostgresStore,
     Row,
     _group_ops,
     _row_to_item,
 )
-from langgraph.store.base import GetOp, ListNamespacesOp, Op, PutOp, Result, SearchOp
 
 logger = logging.getLogger(__name__)
